requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
python-dotenv==1.1.0
